Intent: Consolidar pruebas de glyph_history para clarificar roles de series, contadores y ventanas Operators involved: [Coherence, Self-organization] Affected invariants: [#8, #9] Key changes: - Agrupo la batería de HistoryDict en tests/test_glyph_history_dict.py con docstring diferenciando series y contadores - Reunifico los escenarios de ventanas y helpers en tests/test_glyph_history_windowing.py eliminando duplicados - Documentación integrada sobre diferencias entre series, contadores y ventanas en los nuevos módulos Expected risks/dissonances: Riesgo limitado de cobertura faltante; mitigado replicando los casos previos y ejecutando pytest enfocado Metrics: C(t), Si, νf sin cambios (solo reestructuración de pruebas) Equivalence map: N/A
